fix: Uso del WebService 0117, para obtener registros "cacheados" por EAM

+++ HGT.EAM.WebServices.Infrastructure/Architecture/GridCache/IEamGridFetcher.cs
@@ -9,7 +9,7 @@ public interface IEamGridFetcher
         string username,
         string organization,
         string? password,
-        int gridId,
+        long gridId,
         string gridName,
         string functionName,
         int dataspyId,
